kernel: make 32-bit entropy arithmetic defined

+++ kernel/core/random.c
@@ -35,7 +35,7 @@ static uint64_t arch_entropy_sample(void) {
     v ^= (uint64_t)(uintptr_t)__builtin_return_address(0);
     v ^= (uint64_t)(uintptr_t)&v;
     v ^= arch_read_addr_space_token();
-    v ^= frame_free_count() << 32;
+    v ^= (uint64_t)frame_free_count() << 32;
     v ^= (uint64_t)(uintptr_t)proc_current();
 
     uint64_t hw = 0;
